Struct tor_interface::tor_provider::OnionStream

pub struct OnionStream { /* private fields */ }

A wrapper around a std::net::TcpStream with some Tor-specific customisations

An onion-listener can be constructed using the TorProvider::connect() method.

Implementations

impl OnionStream

pub fn peer_addr(&self) -> Option<TargetAddr>

Returns the target address of the remote peer of this onion connection.

pub fn local_addr(&self) -> Option<OnionAddr>

Returns the onion address of the local connection for an incoming onion-service connection. Returns None for outgoing connections.

pub fn try_clone(&self) -> Result<Self, Error>

Tries to clone the underlying connection and data. A simple pass-through to std::net::TcpStream::try_clone().