Module tor_interface::tor_crypto

Structs

Ed25519PrivateKey

Ed25519PublicKey

Ed25519Signature

V3OnionServiceId

X25519PrivateKey

X25519PublicKey

Enums

Error

SignBit

Constants

ED25519_PRIVATE_KEYBLOB_BASE64_LENGTH

The number of bytes needed to store base64 encoded ed25519 private key as an ASCII c-string (not including null-terminator)

ED25519_PRIVATE_KEYBLOB_HEADER_LENGTH

The number of bytes needed to store the keyblob header

ED25519_PRIVATE_KEYBLOB_LENGTH

The number of bytes needed to store ed25519 private keyblob as an ASCII c-string (not including a null terminator)

ED25519_PRIVATE_KEYBLOB_SIZE

The number of bytes needed to store ed25519 private keyblob as an ASCII c-string (including a null terminator)

ED25519_PRIVATE_KEY_SIZE

The number of bytes in an ed25519 secret key cbindgen:ignore

ED25519_PUBLIC_KEY_SIZE

The number of bytes in an ed25519 public key cbindgen:ignore

ED25519_SIGNATURE_SIZE

The number of bytes in an ed25519 signature cbindgen:ignore

V3_ONION_SERVICE_ID_LENGTH

The number of bytes needed to store onion service id as an ASCII c-string (not including null-terminator)

V3_ONION_SERVICE_ID_SIZE

The number of bytes needed to store onion service id as an ASCII c-string (including null-terminator)

X25519_PRIVATE_KEYBLOB_BASE64_LENGTH

The number of bytes needed to store base64 encoded x25519 private key as an ASCII c-string (not including null-terminator)

X25519_PRIVATE_KEYBLOB_BASE64_SIZE

The number of bytes needed to store base64 encoded x25519 private key as an ASCII c-string (including a null terminator)

X25519_PRIVATE_KEY_SIZE

The number of bytes in an x25519 private key cbindgen:ignore

X25519_PUBLIC_KEYBLOB_BASE32_LENGTH

The number of bytes needed to store base32 encoded x25519 public key as an ASCII c-string (not including null-terminator)

X25519_PUBLIC_KEYBLOB_BASE32_SIZE

The number of bytes needed to store bsae32 encoded x25519 public key as an ASCII c-string (including a null terminator)

X25519_PUBLIC_KEY_SIZE

The number of bytes in an x25519 publickey cbindgen:ignore

Functions

hash_tor_password